+++ src/main/java/org/synyx/urlaubsverwaltung/availability/api/AvailabilityListDto.java
@@ -5,23 +5,19 @@ import java.util.List;
 @Deprecated(forRemoval = true, since = "4.4.0")
 class AvailabilityListDto {
 
-    private final Integer personId;
+    private final Long personId;
     private final List<DayAvailability> availabilities;
 
-    AvailabilityListDto(List<DayAvailability> availabilities, Integer personId) {
-
+    AvailabilityListDto(List<DayAvailability> availabilities, Long personId) {
         this.availabilities = availabilities;
         this.personId = personId;
     }
 
     public List<DayAvailability> getAvailabilities() {
-
         return availabilities;
     }
 
-
-    public Integer getPersonId() {
-
+    public Long getPersonId() {
         return personId;
     }
 }
