+++ src/main/java/org/synyx/urlaubsverwaltung/overtime/OvertimeCommentRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
  *
  * @since 2.11.0
  */
-interface OvertimeCommentRepository extends CrudRepository<OvertimeComment, Integer> {
+interface OvertimeCommentRepository extends CrudRepository<OvertimeComment, Long> {
 
     List<OvertimeComment> findByOvertime(Overtime overtime);
